Adding initial logic for Windows
